update readme and example
