--- examples/python/echo/config.json
@@ -1,8 +1,0 @@
-{
-    "version": "1.0",
-    "runtime": "python",
-    "name": "echo",
-    "formula": "echo.py",
-    "help": "echo.md",
-    "dependencies": "requirements.txt"
-}
