--- examples/python/echo/echo.md
@@ -1,3 +1,0 @@
-# ECHO(value)
-
-Echos back a value into a cell.
