--- examples/python/echo/echo.py
@@ -1,6 +1,0 @@
-from enum import Enum
-from typing import Any
-from sheetrocks import value as v
-
-def calculate(arg1 : Any) -> Any:
-    return arg1
