--- examples/python/echo/requirements.txt
@@ -1,1 +1,0 @@
-# include anything you'd like here!
