--- examples/python/sum/customsum.md
@@ -1,5 +1,0 @@
-# CUSTOMSUM(arg1, arg2, ...args)
-
-Sums all the arguments in the formula.
-
-If an argument is not a number, it ignores the value. Also accepts ranges such as A1:C3.
