--- examples/python/sum/requirements.txt
@@ -1,1 +1,0 @@
-# Include anything you'd like here!
