+++ help_template.md
@@ -1,0 +1,61 @@
+# FORMULANAME
+
+Returns [description of the formula output].
+
+---
+## Syntax
+
+FORMULANAME(arg_1, arg_2, arg_3, [optional_arg])
+
+#### arg_1
+Description of `arg_1`.
+    - Other relevant information about `arg_1`.
+    - Other relevant information about `arg_1`.
+
+#### arg_2
+Description of `arg_2`.
+
+#### arg_3
+Description of `arg_3`.
+
+#### [optional_arg]
+[OPTIONAL - defaults to `DEFAULT_VALUE`] Description of `optional_arg`.
+    - Other relevant information about `optional_arg`.
+
+---
+## Examples
+
+Relevant information about examples given.
+
+EXAMPLE_RANGE
+1,          2,         3
+"test", "string", 11/03/1983
+
+FORMULANAME(A1:B2, 2, 3)=5
+
+FORMULANAME(4, 5, 6)=7
+
+EXAMPLE_RANGE
+1, 2
+
+FORMULANAME(4, 5, 6, TRUE)=4
+
+---
+## Notes
+
+- Bulleted list of useful information about formula.
+- Bulleted list of useful information about formula.
+- Bulleted list of useful information about formula.
+
+---
+## Related Formulas
+
+- [OTHER.FORMULA.NAME](url): Concise description of formula and how it relates to FORMULANAME.
+- [OTHER.FORMULA.NAME](url): Concise description of formula and how it relates to FORMULANAME.
+- [OTHER.FORMULA.NAME](url): Concise description of formula and how it relates to FORMULANAME.
+
+
+
+
+
+
